+++ api/.vscode/tasks.json
@@ -10,7 +10,10 @@
 				"/consoleloggerparameters:NoSummary"
 			],
 			"type": "process",
-			"problemMatcher": "$msCompile"
+			"problemMatcher": "$msCompile",
+			"options": {
+				"cwd": "${workspaceFolder}/"
+			}
 		},
 		{
 			"label": "build",
@@ -26,7 +29,10 @@
 				"kind": "build",
 				"isDefault": true
 			},
-			"problemMatcher": "$msCompile"
+			"problemMatcher": "$msCompile",
+			"options": {
+				"cwd": "${workspaceFolder}/"
+			}
 		},
 		{
 			"label": "clean release",
@@ -39,7 +45,10 @@
 				"/consoleloggerparameters:NoSummary"
 			],
 			"type": "process",
-			"problemMatcher": "$msCompile"
+			"problemMatcher": "$msCompile",
+			"options": {
+				"cwd": "${workspaceFolder}/"
+			}
 		},
 		{
 			"label": "publish",
@@ -53,13 +62,16 @@
 			],
 			"type": "process",
 			"dependsOn": "clean release",
-			"problemMatcher": "$msCompile"
+			"problemMatcher": "$msCompile",
+			"options": {
+				"cwd": "${workspaceFolder}/"
+			}
 		},
 		{
 			"type": "func",
 			"dependsOn": "build",
 			"options": {
-				"cwd": "${workspaceFolder}/bin/Debug/net5.0"
+				"cwd": "${workspaceFolder}/bin/Debug/netcoreapp3.1"
 			},
 			"command": "host start",
 			"isBackground": true,
