+++ api/CountFunction.cs
@@ -1,30 +1,26 @@
-using System.Net;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.Azure.WebJobs;
+using Microsoft.Azure.WebJobs.Extensions.Http;
 using System.Threading.Tasks;
-using Microsoft.Azure.Functions.Worker;
-using Microsoft.Azure.Functions.Worker.Http;
-using Microsoft.Extensions.Logging;
+using WhoIsWho.Portal.Api.Services;
 
 namespace WhoIsWho.Api
 {
     public class CountFunction
     {
-        private readonly services.CognitiveSearchService searchService;
+        private readonly CognitiveSearchService searchService;
 
-        public CountFunction(services.CognitiveSearchService s)
+        public CountFunction(CognitiveSearchService s)
         {
             searchService = s;
         }
 
-        [Function("Count")]
-        public async Task<HttpResponseData> Run([HttpTrigger(AuthorizationLevel.Anonymous, "get")] HttpRequestData req, FunctionContext executionContext)
+        [FunctionName("Count")]
+        public async Task<IActionResult> Run([HttpTrigger(AuthorizationLevel.Anonymous, "get")] HttpRequest req)
         {
-            var json= await searchService.Count();
-
-            var response = req.CreateResponse(HttpStatusCode.OK);
-            response.Headers.Add("Content-Type", "application/json; odata.metadata=minimal");
-            response.WriteString(json);
-
-            return response;
+            var json = await searchService.Count();
+            return new JsonResult(json);
         }
     }
 }
