+++ api/Properties/serviceDependencies.local.json
