+++ api/ResultByFiltersFunction.cs
@@ -1,33 +1,33 @@
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.Azure.WebJobs;
+using Microsoft.Azure.WebJobs.Extensions.Http;
 using System.Net;
+using System.Net.Http;
+using System.Text;
 using System.Threading.Tasks;
-using Microsoft.Azure.Functions.Worker;
-using Microsoft.Azure.Functions.Worker.Http;
-using Microsoft.Extensions.Logging;
+using WhoIsWho.Portal.Api.Services;
 
 namespace WhoIsWho.Api
 {
     public class ResultByFilterFunction
     {
-        private readonly services.CognitiveSearchService searchService;
+        private readonly CognitiveSearchService searchService;
 
-        public ResultByFilterFunction(services.CognitiveSearchService s)
+        public ResultByFilterFunction(CognitiveSearchService s)
         {
             searchService = s;
         }
 
-        [Function("ResultByFilter")]
-        public async Task<HttpResponseData> Run([HttpTrigger(AuthorizationLevel.Anonymous, "get")] HttpRequestData req, FunctionContext executionContext)
+        [FunctionName("ResultByFilter")]
+        public async Task<HttpResponseMessage> Run([HttpTrigger(AuthorizationLevel.Anonymous, "get")] HttpRequest req)
         {
-            var queryDictionary = System.Web.HttpUtility.ParseQueryString(req.Url.Query);
-            var searchString = queryDictionary["$filter"];
-
-            var json= await searchService.ResultByFilter(searchString);
-
-            var response = req.CreateResponse(HttpStatusCode.OK);
-            response.Headers.Add("Content-Type", "application/json; odata.metadata=minimal");
-            response.WriteString(json);
-
-            return response;
+            var searchString = req.Query["$filter"];
+            var json = await searchService.ResultByFilter(searchString);
+            return new HttpResponseMessage(HttpStatusCode.OK)
+            {
+                Content = new StringContent(json, Encoding.UTF8, "application/json")
+            };
         }
     }
 }
