--- api/WhoIsWho.Portal.API.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31129.286
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WhoIsWho.Portal.API", "WhoIsWho.Portal.API\WhoIsWho.Portal.API.csproj", "{81E041D2-271D-4180-B1C6-11D2E97B4F54}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{81E041D2-271D-4180-B1C6-11D2E97B4F54}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{81E041D2-271D-4180-B1C6-11D2E97B4F54}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{81E041D2-271D-4180-B1C6-11D2E97B4F54}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{81E041D2-271D-4180-B1C6-11D2E97B4F54}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {E44B1564-1CCC-41DF-9D5F-24202A012A4D}
-	EndGlobalSection
-EndGlobal
