--- api/WhoIsWho.Portal.API/local.settings.sample.json
@@ -1,12 +1,0 @@
-{
-  "IsEncrypted": false,
-  "Host": {
-    "CORS": "*"
-  },
-  "Values": {
-    "AzureWebJobsStorage": "UseDevelopmentStorage=true",
-    "FUNCTIONS_WORKER_RUNTIME": "dotnet",
-    "SEARCH_SERVICE_KEY": "",
-    "SEARCH_SERVICE_URL": ""
-  }
-}
