+++ api/local.settings.sample.json
@@ -4,8 +4,8 @@
     "CORS": "*"
   },
   "Values": {
-    "AzureWebJobsStorage": "",
-    "FUNCTIONS_WORKER_RUNTIME": "dotnet-isolated",
+    "AzureWebJobsStorage": "UseDevelopmentStorage=true",
+    "FUNCTIONS_WORKER_RUNTIME": "dotnet",
     "APPINSIGHTS_INSTRUMENTATIONKEY": "",
     "SEARCH_SERVICE_KEY": "",
     "SEARCH_SERVICE_URL": ""
